+++ docs/examples/imu_dot.py
@@ -1,11 +1,12 @@
+from __future__ import division  # for py2.x compatibility
 from pisense import SenseHAT, array
 from colorzero import Color
 from time import sleep
 
 def movements(imu):
     for reading in imu:
-        delta_x = round(max(-1, min(1, imu.accel.x)))
-        delta_y = round(max(-1, min(1, imu.accel.y)))
+        delta_x = int(round(max(-1, min(1, imu.accel.x))))
+        delta_y = int(round(max(-1, min(1, imu.accel.y))))
         if delta_x != 0 or delta_y != 0:
             yield delta_x, delta_y
         sleep(1/10)
